`compute_classification_report` --> `MinervaTask`
